Change in character size on button label
You can estimate the number of characters the button can contain using its size and text size and then limit the size of the label.
Looking at the drawButton function I see that it positions the cursor using a calculation to center the string, it is "(_w / 2) - (strlen (_label) * 3 * _textsize_x)".
From there I have obtained an equation that estimates the number of characters: "n = w / (6 * textsize_x)". I have taken into account the division by 0 taking that value to 1 (being the smallest text size).
I also observe that the label is defined as a vector "char _label [10]" so I have defined it as "char * _label" and dynamically creating the memory with the new operator after obtaining the estimated number of characters for that button.
I've tried changing text size, button size and with different label sizes and the text remains inside the button, everything looks great.